requirements.txt
torch>=2.0.0
numpy>=1.24.0
rasterio>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
Pillow>=9.5.0
scipy>=1.10.0
pyproj>=3.5.0
requests>=2.28.0
faiss-cpu>=1.7.0
